+++ lib/core/lib/generators/ros/endpoint/endpoint_generator.rb
@@ -7,22 +7,6 @@ module Ros
   class EndpointGenerator < Rails::Generators::NamedBase
     include GeneratorsHelper
 
-    source_root File.expand_path('templates', __dir__)
-
-    def create_files
-      generate_resource
-      generate_request_spec
-      generate_resource_spec
-      generate_policy
-      generate_policy_spec
-      generate_api_doc
-      generate_model
-      generate_route
-      generate_controller
-    end
-
-    private
-
     def generate_request_spec
       Ros::RequestSpecGenerator.new([name]).invoke_all
     end
@@ -62,6 +46,10 @@ module Ros
 
     def generate_model
       Ros::ModelGenerator.new([name]).invoke_all
+    end
+
+    def generate_model_specs
+      Ros::ModelSpecGenerator.new([name]).invoke_all
     end
   end
 end
