+++ lib/core/lib/generators/ros/model_spec_generator/model_spec_generator.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Ros
+  class ModelSpecGenerator < Rails::Generators::NamedBase
+    def create_files
+      insert_into_file "spec/models/#{name}_spec.rb", after: ":model do\n" do
+        "  include_examples 'application record concern' do\n"\
+        "    let(:tenant) { Tenant.first }\n"\
+        "    let!(:subject) { create(factory_name) }\n"\
+        "  end\n\n"
+      end
+    end
+  end
+end
